func : Creacion funcion para abrir y cerra modal cuando se envio el formulario
